requirements.txt
Flask==2.3.0
Flask-SocketIO==5.3.0
gunicorn==21.2.0
eventlet==0.33.3
mysql-connector-python==8.2.0
Flask-Failsafe==0.2
cryptography==41.0.0